miner: move TransactionByPriceAndNonce to miner package
